Add a backup sound to potentially use when the robot is backing up.
Unfortunately WPIlib doesn't have a speaker thing, but I'm sure someone
could figure out how to hook up a speaker as an analog out and play the
sound through that in parallel with all of the other stuff. I think the
cRIO doesn't have the oomph to deal with both playing a sound and doing
motor stuff; we get errors as it is right now.